prescription edit: refactor
